fix : api post and set value fixed
