+++ README.md
@@ -3,3 +3,5 @@ Frontend skeleton with ionic + antd + styled components.
 
 # Serve Application
 $ ionic serve
+
+npm install --legacy-peer-deps
